@@ -15,72 +15,72 @@
     "patch:hosts": "fec patch-etc-hosts",
     "start": "fec dev",
     "test": "TZ=UTC jest --verbose --no-cache",
-    "verify": "npm-run-all build lint test"
+    "verify": "npm-run-all build lint test",
+    "postinstall": "ts-patch install"
   },
   "dependencies": {
-    "@patternfly/react-core": "^4.206.2",
-    "@patternfly/react-icons": "^4.53.16",
-    "@patternfly/react-table": "^4.75.2",
-    "@redhat-cloud-services/frontend-components": "^3.9.22",
-    "@redhat-cloud-services/frontend-components-config": "^4.6.34",
-    "@redhat-cloud-services/frontend-components-notifications": "^3.2.12",
-    "@redhat-cloud-services/frontend-components-utilities": "^3.3.5",
-    "@redhat-cloud-services/insights-common-typescript": "^4.12.3",
-    "@testing-library/dom": "^8.13.0",
-    "@types/jest": "^27.5.0",
-    "@types/node": "^17.0.31",
-    "@types/react": "^18.0.8",
-    "@types/react-dom": "^18.0.3",
-    "@unleash/proxy-client-react": "^3.0.0",
+    "@patternfly/react-core": "^5.0.0",
+    "@patternfly/react-icons": "^5.0.0",
+    "@patternfly/react-table": "^5.0.0",
+    "@redhat-cloud-services/frontend-components": "^4.0.0",
+    "@redhat-cloud-services/frontend-components-notifications": "^4.0.0",
+    "@redhat-cloud-services/frontend-components-utilities": "^4.0.0",
+    "@redhat-cloud-services/rbac-client": "^1.2.8",
+    "@unleash/proxy-client-react": "^3.6.0",
     "axios": "^0.21.1",
+    "classnames": "^2.3.2",
     "dayjs": "^1.11.7",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "formik": "^2.2.9",
+    "formik": "^2.4.5",
     "lodash": "^4.17.21",
-    "prettier": "^2.7.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-json-view": "^1.21.3",
-    "react-jss": "^10.9.0",
-    "react-query": "^3.39.1",
-    "react-redux": "^8.0.1",
-    "react-router-dom": "^6.0.0",
-    "redux": "^4.2.0",
-    "typescript": "^4.6.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-jss": "^10.10.0",
+    "react-query": "^3.39.3",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.2",
+    "react18-json-view": "^0.2.5",
+    "redux": "^4.2.1",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/plugin-proposal-class-properties": "^7.16.7",
-    "@babel/plugin-proposal-object-rest-spread": "^7.17.3",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
-    "@babel/preset-react": "^7.16.7",
-    "@babel/preset-typescript": "^7.16.7",
-    "@testing-library/dom": "^8.13.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/react-hooks": "^7.0.0",
-    "@testing-library/user-event": "14.1.1",
-    "@types/jest": "^27.5.0",
-    "@types/lodash": "^4.14.182",
+    "@babel/core": "^7.22.20",
+    "@babel/plugin-transform-object-rest-spread": "^7.22.15",
+    "@babel/plugin-transform-runtime": "^7.22.15",
+    "@babel/preset-env": "^7.22.20",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/preset-typescript": "^7.22.15",
+    "@redhat-cloud-services/eslint-config-redhat-cloud-services": "^1.2.6",
+    "@redhat-cloud-services/frontend-components-config": "^6.0.5",
+    "@redhat-cloud-services/tsc-transform-imports": "^1.0.4",
+    "@redhat-cloud-services/types": "^1.0.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.0",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^20.6.3",
+    "@types/react": "^18.2.22",
+    "@types/react-dom": "^18.2.5",
     "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.29.0",
-    "@typescript-eslint/parser": "^5.29.0",
-    "babel-jest": "^28.0.3",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "babel-jest": "^29.7.0",
     "babel-plugin-transform-imports": "^2.0.0",
-    "eslint": "^7.0.0",
-    "eslint-plugin-react": "^7.30.0",
+    "eslint": "^8.35.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "identity-obj-proxy": "^3.0.0",
-    "insights-common-typescript-dev": "^0.1.21",
-    "jest": "^28.0.3",
-    "jest-environment-jsdom": "^28.0.3",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "npm-run-all": "^4.1.5",
-    "ts-jest": "^28.0.1",
-    "webpack-bundle-analyzer": "^4.5.0"
+    "prettier": "^3.0.3",
+    "prop-types": "15.8.0",
+    "ts-jest": "^29.1.1",
+    "ts-patch": "^3.0.2",
+    "typescript": "^5.0.0",
+    "webpack-bundle-analyzer": "^4.9.1"
   },
   "insights": {
     "appname": "content-sources",
