@@ -1,7 +1,8 @@
-import ReactDOM from 'react-dom';
-
+import { createRoot } from 'react-dom/client';
 import AppEntry from './AppEntry';
 
-const root = document.getElementById('root');
 
-ReactDOM.render(<AppEntry />, root, () => root?.setAttribute('data-ouia-safe', 'true'));
+const container = document.getElementById('app');
+const root = createRoot(container!)
+
+root.render(<AppEntry />);
