@@ -1,25 +1,27 @@
 require.extensions['.css'] = () => undefined;
 
 module.exports = {
-  presets: [['@babel/env', { modules: 'commonjs' }], '@babel/react'],
-  plugins: [
-    '@babel/plugin-transform-runtime',
-    '@babel/plugin-syntax-dynamic-import',
-    '@babel/plugin-proposal-object-rest-spread',
-    '@babel/plugin-proposal-class-properties',
-    [
-      'transform-imports',
-      {
-        '@patternfly/react-icons': {
-          transform: (importName) =>
-            `@patternfly/react-icons/dist/js/icons/${importName
-              .split(/(?=[A-Z])/)
-              .join('-')
-              .toLowerCase()}`,
-          preventFullImport: true,
-        },
-      },
-      'react-icons',
+    presets: [
+        ['@babel/preset-env', { modules: 'commonjs' }],
+        '@babel/react',
+        '@babel/preset-typescript'
     ],
-  ],
+    plugins: [
+        '@babel/plugin-proposal-object-rest-spread',
+        '@babel/plugin-transform-runtime',
+        [
+            'transform-imports',
+            {
+                '@patternfly/react-icons': {
+                    transform: (importName) =>
+                        `@patternfly/react-icons/dist/js/icons/${importName
+                            .split(/(?=[A-Z])/)
+                            .join('-')
+                            .toLowerCase()}`,
+                    preventFullImport: true,
+                },
+            },
+            'react-icons',
+        ],
+    ],
 };
