@@ -1,6 +1,6 @@
-import { mockInsights } from 'insights-common-typescript-dev';
+// import { mockInsights } from 'insights-common-typescript-dev';
 import React from 'react';
 
 declare const global;
 global.React = React;
-mockInsights();
+// mockInsights();
