@@ -13,7 +13,11 @@ module.exports = {
    */
   plugins: [],
   moduleFederation: {
-    exclude: ['react-router-dom'],
-    shared: [{ 'react-router-dom': { singleton: true } }],
+    exclude: [
+        // 'react-router-dom'
+    ],
+    shared: [
+        // { 'react-router-dom': { singleton: true } }
+    ],
   },
 };
