@@ -2,7 +2,7 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'jsdom',
   coverageDirectory: './coverage',
-  collectCoverageFrom: ['src/**/*.{ts,tsx}', '!**/node_modules/**', '!test/**'],
+  collectCoverageFrom: ['src/**/*.test.{ts,tsx}', '!**/node_modules/**', '!test/**'],
   coveragePathIgnorePatterns: [],
   setupFiles: ['<rootDir>/config/setupTests.ts'],
   setupFilesAfterEnv: ['<rootDir>/config/setupTestFramework.ts'],
@@ -11,4 +11,5 @@ module.exports = {
     '\\.(svg)$': 'identity-obj-proxy',
     '\\.(css|scss)$': 'identity-obj-proxy',
   },
+  transformIgnorePatterns: [ '<rootDir>/node_modules/(?!@redhat-cloud-services)' ],
 };
