@@ -82,19 +82,25 @@ it('mapValidationData', () => {
 
 it('Notifies on file upload failure due to size', () => {
   const notif = jest.fn((payload) => payload);
-  const f = new File([''], 'filename', { type: 'text/html' });
-  Object.defineProperty(f, 'size', { value: maxUploadSize + 1 });
+  const file = new File([''], 'filename', { type: 'text/html' });
+  Object.defineProperty(file, 'size', { value: maxUploadSize + 1 });
 
-  failedFileUpload([f], notif);
+  failedFileUpload([{
+    file,
+    errors: []
+  }], notif);
   expect(notif.mock.calls).toHaveLength(1);
   expect(notif.mock.calls[0][0].description).toMatch(/file is larger than/);
 });
 
 it('Notifies on file upload failure due to too many files', () => {
   const notif = jest.fn((payload) => payload);
-  const f = new File([''], 'filename', { type: 'text/html' });
+  const file = new File([''], 'filename', { type: 'text/html' });
 
-  failedFileUpload([f, f], notif);
+  failedFileUpload([{
+    file,
+    errors: []
+  }], notif);
   expect(notif.mock.calls).toHaveLength(1);
   expect(notif.mock.calls[0][0].description).toMatch(/Only a single file upload is supported/);
 });
