@@ -2,9 +2,10 @@ import {
   EmptyStateBody,
   EmptyState,
   EmptyStateVariant,
-  Title,
   EmptyStateIcon,
-  Button,
+  Button, 
+  EmptyStateHeader, 
+  EmptyStateFooter,
 } from '@patternfly/react-core';
 import { SearchIcon, PlusCircleIcon } from '@patternfly/react-icons/dist/js/icons';
 import { createUseStyles } from 'react-jss';
@@ -37,13 +38,10 @@ const EmptyTableState = ({
   const classes = useStyles();
   return (
     <EmptyState variant={EmptyStateVariant.full} className={classes.emptyStateContainer}>
-      <EmptyStateIcon icon={notFiltered ? PlusCircleIcon : SearchIcon} />
-      <Title headingLevel='h2' size='lg' ouiaId='empty_custom_title'>
-        {notFiltered ? `No ${itemName}` : `No ${itemName} match the filter criteria`}
-      </Title>
+      <EmptyStateHeader titleText={<>{notFiltered ? `No ${itemName}` : `No ${itemName} match the filter criteria`}</>} icon={<EmptyStateIcon icon={notFiltered ? PlusCircleIcon : SearchIcon} />} headingLevel='h2' />
       <EmptyStateBody className={classes.emptyStateBody}>
         {notFiltered ? notFilteredBody : 'Clear all filters to show more results'}
-      </EmptyStateBody>
+      </EmptyStateBody><EmptyStateFooter>
       {notFiltered ? (
         notFilteredButton
       ) : (
@@ -51,7 +49,7 @@ const EmptyTableState = ({
           Clear all filters
         </Button>
       )}
-    </EmptyState>
+    </EmptyStateFooter></EmptyState>
   );
 };
 
