@@ -1,21 +1,20 @@
 import { createContext, ReactNode, useContext, useEffect, useState } from 'react';
-import { fetchRBAC, Rbac } from '@redhat-cloud-services/insights-common-typescript';
 import { Features } from '../services/Features/FeatureApi';
 import { notificationsReducer } from '@redhat-cloud-services/frontend-components-notifications/redux';
 import { getRegistry as _getRegistry } from '@redhat-cloud-services/frontend-components-utilities/Registry';
-import { insights } from '../../package.json';
+import PackageJson from '../../package.json';
 import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
 import { useFetchFeaturesQuery } from '../services/Features/FeatureQueries';
+import { fetchRBAC, Rbac } from './RbacUtils';
 
 const getRegistry = _getRegistry as unknown as () => { register: ({ notifications }) => void };
-const { appname } = insights;
+const { appname } = PackageJson.insights;
 
 export interface AppContextInterface {
   rbac?: { read: boolean; write: boolean };
   features: Features | null;
   isFetchingFeatures: boolean;
 }
-
 export const AppContext = createContext({} as AppContextInterface);
 
 export const ContextProvider = ({ children }: { children: ReactNode }) => {
